requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
python-dotenv==1.0.0